requirements.txt
streamlit
supabase
pandas
plotly
google-genai
python-dotenv